+++ admin/google_search_console/class-gsc-mapper.php
@@ -39,7 +39,7 @@ class WPSEO_GSC_Mapper {
 	/**
 	 * If there is no platform, just get the first key out of the array and redirect to it.
 	 *
-	 * @param string $platform
+	 * @param string $platform Platform (desktop, mobile, feature phone).
 	 *
 	 * @return mixed
 	 */
@@ -55,7 +55,7 @@ class WPSEO_GSC_Mapper {
 	/**
 	 * Mapping the platform
 	 *
-	 * @param string $platform
+	 * @param string $platform Platform (desktop, mobile, feature phone).
 	 *
 	 * @return mixed
 	 */
@@ -68,7 +68,7 @@ class WPSEO_GSC_Mapper {
 	/**
 	 * Mapping the given platform by value and return its key
 	 *
-	 * @param string $platform
+	 * @param string $platform Platform (desktop, mobile, feature phone).
 	 *
 	 * @return string
 	 */
@@ -83,11 +83,11 @@ class WPSEO_GSC_Mapper {
 	/**
 	 * Mapping the given category by searching for its key.
 	 *
-	 * @param string $category
+	 * @param string $category Issue type.
 	 *
 	 * @return mixed
 	 */
-	public static function category_to_api( $category) {
+	public static function category_to_api( $category ) {
 		if ( ! empty( $category ) && array_key_exists( $category, self::$categories ) ) {
 			return self::$categories[ $category ];
 		}
@@ -98,7 +98,7 @@ class WPSEO_GSC_Mapper {
 	/**
 	 * Mapping the given category by value and return its key
 	 *
-	 * @param string $category
+	 * @param string $category Issue type.
 	 *
 	 * @return string
 	 */
@@ -109,5 +109,4 @@ class WPSEO_GSC_Mapper {
 
 		return $category;
 	}
-
 }